netlink: stop passing Go pointers to cgo callbacks.
The netlink message code currently creates a callback argument struct,
which is Go allocated memory with pointers to Go allocated memory. In
Go 1.5 we can get away with passing this to a cgo call - it is effectively
safe since the struct is kept alive within the Go function while the three
cgo function calls complete, along with the callbacks (if any).

Post Go 1.5 the pointer passing between Go and C has been further defined
and restricted - a C function may only be passed a pointer to Go allocated
memory if it only uses it until that function returns and the Go allocated
memory does not contain pointers to other Go allocated memory. Additionally,
there is now a dynamic check that panics if it detects inappropriate calls.

To resolve this, register the callback arguments and store a Go reference
to it that can be found via an identifier. This identifier is now passed
to the cgo functions and the callback arguments are located once the Go
callback code is run from the C gateway/trampoline. This follows the
recommendation from:

https://github.com/golang/proposal/blob/master/design/12416-cgo-pointers.md