[backport] avoid processing unrelated vrrp packet
currently VRRP is always sent to receive queue that is processed by the
run loop even if it's not for this VRRP group. Because golang evaluate
switch cases in order, theoretically a VRRP flood may stop failover
because there is always packets to process.

The fix is to drop those packets early before it can interupt run loop.

Change-Id: If1a7084ee87e03a110b05ac30e90e1c40779ef8a