Set noninteractive dpkg mode
Prevent dpkg from asking questions and use the defaults as
the unattended installation process cannot reply.

From dpkg manpage:
```
confdef: If a conffile has been modified and the version in the package
did change, always choose the default action without prompting. If there
is no default action it will stop to ask the user unless --force-confnew
or --force-confold is also been given, in which case it will use that to
decide the final action.
```

Also consolidate two apt-get installs.

Bug: 163134297
Tested: On syllogi/159032 of vsphere-osimage-ubuntu-node

Change-Id: Ifbf6524c295ded7b23a2c8de92587313dc7f04b2