Detect and pin docker.io origin in chroot
This removes the need to keep track of which docker version
is installed in the chroot. This will still require the image
docker and the installed docker to be from the same PPA (to avoid any
potential discrepancies in the on-disk format).

Additionally, this adds empty lines present in scripts within
vsphere-osimage-ubuntu-node so that the diffs are minimal.

Bug: 171061533
See also: gkecl/186033

Change-Id: Ifc34ce7e6898a21bbd8f8024a5900665cd5bd2b2