enables fallback and port for hashing schedulers
for sh and mh, by default, port is not included in hashing. This is
against intuition.

This commit enables sh-port and mh-port option.

Also enables sh-fallback and mh-fallback, which is helpful when l/u
threshold set for a service.

Refer to https://lore.kernel.org/patchwork/patch/387724/ on how the
flags are configured to ipvs module.

Change-Id: I40dafe67b5c0ce03915714979404542afed4e60a