Update `go.mod` versions
`go-1.12` is pretty old now, and there are a number of security
advisories on some dependencies that effectively require `>= go-1.17`.

As far as dependencies, we don't have particular requirements so merely
updating to their latest releases is sufficient.

One minor change is to tests, which are needed due to `libnl-3.5.0`:

> nested attributes now use NLA_F_NESTED bit to indicate a nested
> attribute